Rename from ViewStore to StoreProjection
The name ViewStore was a bit misleading, because this structure has no storage, it doesn't hold any state. Yes, the view relies on it, but all it does is to create a layer of translation between views and models.
The new name, StoreProjection, hopefully will give us a better understanding that this structure is not a Store, but a Projection to THE store (there's only one store in your app). A projection of that store is a way to see the store from a unique perspective, it's still the same store but the viewpoint is whatever that View needs to observe or dispatch to.

The function that lifts store into the store projection was also renamed from the confusing "view(...)" to "projection(...)"